requirement.txt
annotated-types==0.5.0
anyio==3.7.1
azure-core==1.29.4
azure-storage-blob==12.17.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.4
dnspython==2.4.2
docuware-client==0.2.2
exceptiongroup==1.1.3
fastapi==0.70.1
h11==0.14.0
idna==3.4
isodate==0.6.1
motor==3.3.1
passlib==1.7.4
pycparser==2.21
pydantic==1.10.12
pydantic_core==2.6.3
pymongo==4.5.0
pymssql==2.2.9
python-dotenv==0.19.0
python-multipart==0.0.6
requests==2.31.0
six==1.16.0
sniffio==1.3.0
starlette==0.16.0
typing_extensions==4.7.1
uvicorn==0.23.2
python-dotenv==0.19.0
motor==3.3.1
passlib==1.7.4
azure-storage-blob==12.17.0
openai==0.27.7
azure-storage-file-share==12.5.0
azure-ai-formrecognizer==3.2.1
dateparser==1.1.8
scikit-learn==1.3.1
pymssql==2.2.9
llama-index==0.7.8
PyPDF2==3.0.1
ruamel.yaml==0.17.32
openpyxl==3.1.2
python-Levenshtein==0.23.0
xlrd==2.0.1
python-Levenshtein==0.23.0
APScheduler==3.10.4
tornado==6.3.3
httpx==0.25.1